move reverse dns to sysinfo
