+++ sysinfo/reverse_dns.go
@@ -1,4 +1,4 @@
-package utils
+package sysinfo
 
 import (
 	"net"
